# Update prettier to v2.1