Update prettier to v2.1.0

closes #203