@nickmccurdy 2020-08-25
Thanks, this does fix the issue in development! I'm noticing the production build is a megabyte larger, but is 10 times faster than on master. Thoughts on if this is acceptable or if we should stick to similar terser options since production builds are less common?

# master build
```
yarn run v1.22.4
$ PLATFORM=chrome node scripts/build.js
Hash: 9394f6adbd19593723ae
Version: webpack 4.44.1
Time: 132588ms
Built at: 08/25/2020 3:12:48 AM
     Asset       Size  Chunks                    Chunk Names
content.js   7.47 MiB       0  [emitted]  [big]  content
REDACTED
✨  Done in 135.63s.
```

# this build
```
yarn run v1.22.4
$ PLATFORM=chrome node scripts/build.js
Hash: a519ca1d8ac56bf92b9a
Version: webpack 4.44.1
Time: 10306ms
Built at: 08/25/2020 3:10:11 AM
     Asset       Size  Chunks                    Chunk Names
content.js   6.33 MiB       0  [emitted]  [big]  content
REDACTED
✨  Done in 13.84s.
```