@sibiraj-s 2020-08-26
Also, in coming days. if possible. we can move this to background scripts rather than content-scripts. so that we don't need to load them in main pages.