@sibiraj-s 2020-08-26
Actually this build is a smaller than the master build 🤔 

**Master**: content.js   7.47 MiB
**Current Branch**: content.js   6.33 MiB