@nickmccurdy 2020-08-26
@kaicataldo I meant changing the webpack config, not avoiding the upgrade