@kaicataldo 2020-08-27
I believe I added the `TerserPlugin` explicitly so that we could add `output: { ascii_only: true },`. Since this PR fixes the bug in dev, relying on Webpack's defaults for production minification sounds good to me.