@nickmccurdy 2020-09-02
> Actually this build is a smaller than the master build 🤔

I'm getting similar results, I must have mixed them up, sorry! I'll merge this.

> Also, in coming days. if possible. we can move this to background scripts rather than content-scripts. so that we don't need to load them in main pages.

I'm not sure how this would work or what the advantage would be, but feel free to make an issue or pull request if you're still interested in it.