@sibiraj-s 2020-09-03
> I'm not sure how this would work or what the advantage would be

some advantages are.

- background scripts are not loaded in users site.
- background scripts doesn't have strict utf checks.
- background scripts are loaded only once when the extension is installed. (content is loaded from the storage so loading only once doesn't give any particular advantage. just adding a point)

> but feel free to make an issue or pull request if you're still interested in it.

Sure. I will evaluate if this is really helpful and propose something.